typo fix

+++ rlpyt/algos/utils.py
@@ -49,6 +49,7 @@ def discount_return_n_step(reward, done, n_step, discount, return_dest=None,
 
 
 def valid_from_done(done):
-    valid = torch.ones_like(done, dtype=torch.float)
+    done = done.type(torch.float)
+    valid = torch.ones_like(done)
     valid[1:] = 1 - torch.clamp(torch.cumsum(done[:-1], dim=0), max=1)
     return valid
